test/e2e: migrate headless-stateful-serving-cert-secret-delete-data test for OTE compatibility
This commit migrates the headless-stateful-serving-cert-secret-delete-data test to work with
both standard Go testing and OTE (OpenShift Tests Extension) framework.

Changes:
- Add Ginkgo wrapper in e2e.go for OTE test discovery (6 lines)
- Move 5 helper functions from e2e_test.go to e2e.go:
  * pollForServiceServingSecretWithReturn (14 lines)
  * pollForUpdatedServingCert (23 lines)
  * createStatefulSet (64 lines)
  * checkClientPodRcvdUpdatedServerCert (3 lines - wrapper)
  * pollForRunningStatefulSet (30 lines)
- Extract inline test code into shared testHeadlessStatefulServingCertSecretDeleteData
  function (65 lines)
- Change helper function signatures from *testing.T to testing.TB for dual framework compatibility
- Refactor e2e_test.go to call shared function instead of inline code (-218 lines)
- Remove unused imports from e2e_test.go (appsv1, pointer)

Net change: -5 lines (213 added - 218 removed)

This follows the pure code transfer approach - all functions were moved as-is from
e2e_test.go to e2e.go with only signature changes (*testing.T → testing.TB) and
import cleanup. No new logic or code was added.

Test functionality:
This test verifies that when a serving cert secret is deleted for a headless service
backed by a StatefulSet, the cert is regenerated and all StatefulSet pods can use
the new certificate for successful connections.

Both test frameworks continue to work:
- Standard Go test: go test -run "^TestE2E$/^headless-stateful-serving-cert-secret-delete-data$"
- OTE: service-ca-operator-tests-ext (test discovered via Ginkgo)
